requirements.txt
pillow>=10.3.0
numpy==1.26.3
opencv-python==4.9.0.80
requests==2.31.0
pika==1.3.2
python-dotenv==1.0.0
